apps/server/requirements.txt
amqp==5.2.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
billiard==4.2.0
boto3==1.34.105
botocore==1.34.105
celery==5.4.0
certifi==2024.6.2
cffi==1.16.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.0.6
django-cors-headers==4.3.1
django-environ==0.11.2
django-minio-storage==0.5.7
django-s3-file-field==1.0.1
djangorestframework==3.15.1
geojson==3.1.0
jmespath==1.0.1
kombu==5.3.7
minio==7.2.7
mysqlclient==2.2.4
numpy==2.0.0
pandas==2.2.2
prompt_toolkit==3.0.46
pycparser==2.22
pycryptodome==3.20.0
python-dateutil==2.9.0.post0
pytz==2024.1
redis==5.0.5
roifile==2024.5.24
s3transfer==0.10.1
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.1
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
